Struct lowering::lir::BlockGraph

pub struct BlockGraph {
    pub firm: Graph,
    pub blocks: HashMap<Block, Ptr<BasicBlock>>,
    pub vars: HashMap<Node, Var>,
    pub var_counter: usize,
    pub start_block: Ptr<BasicBlock>,
    pub end_block: Ptr<BasicBlock>,
}

A graph of basic blocks.

Fields
firm: Graph blocks: HashMap<Block, Ptr<BasicBlock>> vars: HashMap<Node, Var> var_counter: usize start_block: Ptr<BasicBlock> end_block: Ptr<BasicBlock>
Methods

impl BlockGraph

pub fn iter_blocks<'g>(
&'g self
) -> impl Iterator<Item = Ptr<BasicBlock>> + 'g

Iterate over all basic blocks in self in a breadth-first manner (in control flow direction, starting at the start block)

pub fn postorder_blocks(&self) -> Vec<Ptr<BasicBlock>>

Gives a list of all BasicBlocks in the graph in postorder.

pub fn get_block(&self, firm_block: Block) -> Ptr<BasicBlock>